this works fixed a bug in the padding in the dataset file and siwch the number to 21 for zstd and FP32 for all

+++ CAESAR/models/caesar_decompress.cpp
@@ -170,7 +170,7 @@ torch::Tensor Decompressor::decompress(const unsigned int batch_size,
     }
 
     std::vector<torch::Tensor> hyper_outputs = hyper_decompressor_model_->run(
-        {decoded_hyper_latents.to(torch::kDouble).to(device_)});
+        {decoded_hyper_latents.to(torch::kFloat32).to(device_)});
     torch::Tensor mean = hyper_outputs[0].to(torch::kFloat32);
 
     torch::Tensor latent_indexes_recon = hyper_outputs[1].to(torch::kInt32);
@@ -200,7 +200,7 @@ torch::Tensor Decompressor::decompress(const unsigned int batch_size,
     torch::Tensor reshaped_latents = q_latent_with_offset.reshape(new_shape);
 
     std::vector<torch::Tensor> decompressor_outputs =
-        decompressor_model_->run({reshaped_latents.to(torch::kDouble)});
+        decompressor_model_->run({reshaped_latents.to(torch::kFloat32)});
     torch::Tensor raw_output = decompressor_outputs[0].to(torch::kFloat32);
