+++ CAESAR_hyper_decompressor.py
@@ -406,7 +406,7 @@ model = CompressorMix(
 
 state_dict = remove_module_prefix(torch.load('./pretrained/caesar_v.pt', map_location=device))
 model.load_state_dict(state_dict)
-model = model.double()
+model = model.float()
 
 quantized_cdf, cdf_length, offset = model.entropy_model.prior._update(30)
 medians = model.entropy_model.prior.medians.detach()
@@ -421,7 +421,7 @@ model.eval()
 with torch.no_grad():
     print('device: ', device)
     model = model.to(device)
-    example_inputs=(torch.randn(8, 64, 4, 4, device=device).double(),)
+    example_inputs=(torch.randn(8, 64, 4, 4, device=device).float(),)
     batch_dim = torch.export.Dim("batch", min=1, max=1024)
     # [Optional] Specify the first dimension of the input x as dynamic.
     exported = torch.export.export(model, example_inputs, dynamic_shapes={"x": {0: batch_dim}})
